backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
scikit-learn==1.3.2
pandas==2.1.4
joblib==1.3.2
numpy==1.26.2

sentence-transformers
google-generativeai
python-dotenv
pypdf
supabase
python-multipart
openai-whisper
